+++ src/main/java/com/dddd/doctorpatientrest/application/services/service_impls/FullInfoServiceImpl.java
@@ -59,6 +59,10 @@ public class FullInfoServiceImpl implements FullInfoService {
 
 	@Override
 	public void deleteById(long id) {
+		Optional<FullInfo> fullInfo = fullInfoRepository.findById(id);
+		if (!fullInfo.isPresent()) {
+			throw new ResourceNotFoundException(Constants.FULL_INFO_NOT_FOUND, id);
+		}
 		fullInfoRepository.deleteById(id);
 	}
 }
