+++ src/main/java/com/dddd/doctorpatientrest/application/services/service_impls/PatientServiceImpl.java
@@ -95,6 +95,10 @@ public class PatientServiceImpl implements PatientService {
 
 	@Override
 	public void deleteById(long id) {
+		Optional<Patient> patient = patientRepository.findById(id);
+		if (!patient.isPresent()) {
+			throw new ResourceNotFoundException(Constants.PATIENT_NOT_FOUND, id);
+		}
 		patientRepository.deleteById(id);
 	}
 }
